refactor stratum model methods into files
