+++ client/lib/strata/stratum/model/createGraph.js
@@ -1,0 +1,6 @@
+const graphology = require('graphology')
+
+module.exports = function () {
+  // Create a new graph
+  return new graphology.Graph()
+}
